refactor: integrate config
