+++ unilink/config/config_manager.hpp
@@ -7,7 +7,7 @@
 #include <iostream>
 
 namespace unilink {
-namespace config_manager {
+namespace config {
 
 /**
  * Thread-safe configuration manager implementation
@@ -62,5 +62,5 @@ private:
     std::any deserialize_value(const std::string& value_str, ConfigType type) const;
 };
 
-} // namespace config_manager
+} // namespace config
 } // namespace unilink
