+++ unilink/unilink.hpp
@@ -17,9 +17,9 @@
 #include "unilink/builder/unified_builder.hpp"
 
 // 새로운 Configuration Management API includes
-#include "unilink/config_manager/iconfig_manager.hpp"
-#include "unilink/config_manager/config_manager.hpp"
-#include "unilink/config_manager/config_factory.hpp"
+#include "unilink/config/iconfig_manager.hpp"
+#include "unilink/config/config_manager.hpp"
+#include "unilink/config/config_factory.hpp"
 
 namespace unilink {
 
@@ -105,14 +105,14 @@ inline builder::SerialBuilder serial(const std::string& device, uint32_t baud_ra
 // === 새로운 Configuration Management API ===
 // 편의 별칭들
 namespace config_manager {
-    using IConfigManager = config_manager::IConfigManager;
-    using ConfigManager = config_manager::ConfigManager;
-    using ConfigFactory = config_manager::ConfigFactory;
-    using ConfigPresets = config_manager::ConfigPresets;
-    using ConfigType = config_manager::ConfigType;
-    using ConfigItem = config_manager::ConfigItem;
-    using ValidationResult = config_manager::ValidationResult;
-    using ConfigChangeCallback = config_manager::ConfigChangeCallback;
+    using IConfigManager = config::IConfigManager;
+    using ConfigManager = config::ConfigManager;
+    using ConfigFactory = config::ConfigFactory;
+    using ConfigPresets = config::ConfigPresets;
+    using ConfigType = config::ConfigType;
+    using ConfigItem = config::ConfigItem;
+    using ValidationResult = config::ValidationResult;
+    using ConfigChangeCallback = config::ConfigChangeCallback;
 }
 
 // === 공통 유틸리티 ===
